Upload algolia index on post build
