+++ site/gatsby-site/package.json
@@ -21,7 +21,7 @@
     "@mdx-js/react": "^1.6.22",
     "@philpl/buble": "^0.19.7",
     "@playlyfe/gql": "^2.6.2",
-    "algoliasearch": "^3.35.1",
+    "algoliasearch": "^4.10.5",
     "babel-plugin-styled-components": "^1.12.0",
     "bootstrap": "^4.6.0",
     "bson": "^4.2.3",
@@ -59,7 +59,9 @@
     "gatsby-transformer-sharp": "^3.10.0",
     "graphql": "^15.5.1",
     "is-absolute-url": "^3.0.3",
+    "lodash.chunk": "^4.2.0",
     "lodash.flatten": "^4.4.0",
+    "lodash.isequal": "^4.5.0",
     "lodash.startcase": "^4.4.0",
     "md5": "^2.3.0",
     "react": "^17.0.2",
@@ -70,7 +72,7 @@
     "react-github-btn": "^1.2.0",
     "react-helmet": "^6.1.0",
     "react-id-generator": "^3.0.1",
-    "react-instantsearch-dom": "^6.10.3",
+    "react-instantsearch-dom": "^6.12.1",
     "react-live": "^2.2.3",
     "react-loadable": "^5.5.0",
     "react-markdown": "^6.0.2",
